Separated Map class
