+++ Transity/x64/Debug/Transity.tlog/Cl.items.tlog
@@ -27,5 +27,8 @@ C:\Users\Ziggy\Documents\Transity\Transity\src\modding\Sandbox.cpp;C:\Users\Zigg
 C:\Users\Ziggy\Documents\Transity\Transity\src\core\ThreadManager.cpp;C:\Users\Ziggy\Documents\Transity\Transity\Transity\x64\Debug\ThreadManager.obj
 C:\Users\Ziggy\Documents\Transity\Transity\src\core\ThreadPool.cpp;C:\Users\Ziggy\Documents\Transity\Transity\Transity\x64\Debug\ThreadPool.obj
 C:\Users\Ziggy\Documents\Transity\Transity\src\world\Line.cpp;C:\Users\Ziggy\Documents\Transity\Transity\Transity\x64\Debug\Line.obj
+C:\Users\Ziggy\Documents\Transity\Transity\src\world\managers\CityManager.cpp;C:\Users\Ziggy\Documents\Transity\Transity\Transity\x64\Debug\CityManager.obj
+C:\Users\Ziggy\Documents\Transity\Transity\src\world\managers\LineManager.cpp;C:\Users\Ziggy\Documents\Transity\Transity\Transity\x64\Debug\LineManager.obj
+C:\Users\Ziggy\Documents\Transity\Transity\src\world\managers\TrainManager.cpp;C:\Users\Ziggy\Documents\Transity\Transity\Transity\x64\Debug\TrainManager.obj
 C:\Users\Ziggy\Documents\Transity\Transity\src\world\Map.cpp;C:\Users\Ziggy\Documents\Transity\Transity\Transity\x64\Debug\Map.obj
 C:\Users\Ziggy\Documents\Transity\Transity\src\world\managers\SelectionManager.cpp;C:\Users\Ziggy\Documents\Transity\Transity\Transity\x64\Debug\SelectionManager.obj
