+++ Transity/x64/Debug/Transity.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\CAMERA.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\DEBUG.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\ERRORHANDLER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\GAME.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\GAMESETTINGS.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\IMGUI-SFML.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\IMGUI.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\IMGUI_DRAW.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\IMGUI_TABLES.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\IMGUI_WIDGETS.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\INPUTMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\LINE.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\MAP.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\PASSENGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\PLUGINLOADER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\PLUGINMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\PROFILER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\RENDERER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\RESOURCEMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\SANDBOX.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\SAVEMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\SELECTIONMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\SETTINGSREGISTRY.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\SIMULATION.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\STATEMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\THREADMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\THREADPOOL.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\TRAIN.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\TRANSITY.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\UIMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\WINDOWMANAGER.OBJ
+^C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\CAMERA.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\CITYMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\DEBUG.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\ERRORHANDLER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\GAME.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\GAMESETTINGS.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\IMGUI-SFML.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\IMGUI.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\IMGUI_DRAW.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\IMGUI_TABLES.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\IMGUI_WIDGETS.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\INPUTMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\LINE.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\LINEMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\MAP.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\PASSENGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\PLUGINLOADER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\PLUGINMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\PROFILER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\RENDERER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\RESOURCEMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\SANDBOX.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\SAVEMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\SELECTIONMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\SETTINGSREGISTRY.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\SIMULATION.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\STATEMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\THREADMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\THREADPOOL.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\TRAIN.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\TRAINMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\TRANSITY.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\UIMANAGER.OBJ|C:\USERS\ZIGGY\DOCUMENTS\TRANSITY\TRANSITY\TRANSITY\X64\DEBUG\WINDOWMANAGER.OBJ
 C:\Users\Ziggy\Documents\Transity\Transity\Transity\x64\Debug\Transity.ilk
