+++ src/world/Map.h
@@ -13,6 +13,9 @@
 #include "Segment.h"
 #include "managers/HandleManager.h"
 #include "managers/SelectionManager.h"
+#include "managers/CityManager.h"
+#include "managers/LineManager.h"
+#include "managers/TrainManager.h"
 #include "../core/StateManager.h"
 #include "../core/InputManager.h"
 #include "../Constants.h"
@@ -31,9 +34,12 @@ public:
         m_minRadius(100),
         stateManager(sm),
         startCityForTrain(nullptr), endCityForTrain(nullptr),
-        selectionManager(*this) {}
+        selectionManager(*this), cityManager(m_minRadius, *this), lineManager(*this), trainManager(*this, stateManager) {}
 
     SelectionManager& GetSelectionManager() { return selectionManager; }
+    CityManager& GetCityManager() { return cityManager; }
+    LineManager& GetLineManager() { return lineManager; }
+    TrainManager& GetTrainManager() { return trainManager; }
     const SelectionManager& GetSelectionManager() const { return selectionManager; }
 
     // Tile management
@@ -52,39 +58,36 @@ public:
     std::list<Node>& GetNodes() { return m_nodes; }
 
     // City management
-    void AddCity(const sf::Vector2f pos);
-    std::list<City>& GetCities() { return m_cities; }
-    void RemoveCity();
-    void MoveCity(sf::Vector2f newPos);
+    std::list<City>& GetCities() { return cityManager.GetCities(); }
 
     // Line management
-    void UseLineMode(sf::Vector2f pos);
-    void CreateLine(sf::Vector2f pos);
-    void AddToLineStart(sf::Vector2f pos);
-    void AddToLineEnd(sf::Vector2f pos);
-    void RemoveLine();
-    std::list<Line>& GetLines() { return m_lines; }
-    void MoveSelectedLineHandle(sf::Vector2f newPos);
-    void UpdateSharedSegments();
+    std::list<Line>& GetLines() { return lineManager.GetLines(); }
     std::vector<Segment> GetSharedSegments() const { return sharedSegments; }
     bool isLineSelected() { return selectionManager.GetSelectedLine() == nullptr ? false : true; }
     std::vector<Node*> FindRouteBetweenNodes(Node* start, Node* end);
 
     // Train management
-    void UseTrainPlaceMode(sf::Vector2f pos, bool left);
-    void AddTrain();
-    void RemoveTrain();
-    std::vector<std::unique_ptr<Train>>& GetTrains() { return m_trains; }
+    std::vector<std::unique_ptr<Train>>& GetTrains() { return trainManager.GetTrains(); }
     City* GetStartCityForTrain() const { return startCityForTrain; }
     City* GetEndCityForTrain() const { return endCityForTrain; }
 
     // Passenger management
     int GetScore() const { return m_score; }
-    void UpdatePassengers(float dt);
+    void SetScore(int value) { m_score = value; }
     void SpawnPassenger(City* origin, City* destination);
 
     // Helper methods
     float DistancePointToSegment(const sf::Vector2f& point, const sf::Vector2f& segStart, const sf::Vector2f& segEnd);
+
+    // Helper functions
+    void Resize(unsigned int newSize);
+    City* FindCityAtPosition(sf::Vector2f pos);
+    std::pair<int, int> NormalizeSegment(int startIndex, int endIndex) const;
+    std::string GenerateSegmentKey(const sf::Vector2f& start, const sf::Vector2f& end) const;
+    bool ArePositionsEqual(const sf::Vector2f& pos1, const sf::Vector2f& pos2, float epsilon = 0.1f);
+    Node* FindGenericNodeAtPosition(sf::Vector2f pos);
+    bool WouldCauseParallelConflict(const sf::Vector2f& segStart, const sf::Vector2f& segEnd);
+    void CreateBranch(Line* parentLine, int branchHandleIndex, sf::Vector2f pos);
 
 private:
     // Grid representation
@@ -94,20 +97,7 @@ private:
     unsigned int m_minRadius;
     int m_score = 0;
 
-    // Helper functions
-    void Resize(unsigned int newSize);
-    City* FindCityAtPosition(sf::Vector2f pos);
-    std::pair<int, int> NormalizeSegment(int startIndex, int endIndex) const;
-    std::string GenerateSegmentKey(const sf::Vector2f& start, const sf::Vector2f& end) const;
-    static bool ArePositionsEqual(const sf::Vector2f& pos1, const sf::Vector2f& pos2, float epsilon = 0.1f);
-    Node* FindGenericNodeAtPosition(sf::Vector2f pos);
-    bool WouldCauseParallelConflict(const sf::Vector2f& segStart, const sf::Vector2f& segEnd);
-    void CreateBranch(Line* parentLine, int branchHandleIndex, sf::Vector2f pos);
-
     // Collections
-    std::list<City> m_cities;
-    std::list<Line> m_lines;
-    std::vector<std::unique_ptr<Train>> m_trains;
     std::list<Node> m_nodes;
 
     std::vector<Segment> sharedSegments; // List of shared segments
@@ -116,5 +106,8 @@ private:
 
     // Selection Manager
     SelectionManager selectionManager;
+    CityManager cityManager;
     StateManager& stateManager;
+    LineManager lineManager;
+    TrainManager trainManager;
 };
